# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "app"
        ]
      }
    },
    "/users": {
      "get": {
        "operationId": "UsersController_getUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all users",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "User already exists"
          }
        },
        "summary": "Create a new user",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_deleteUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User deleted"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Delete a user",
        "tags": [
          "users"
        ]
      }
    },
    "/users/validate": {
      "post": {
        "operationId": "UsersController_validate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Validate user credentials",
        "tags": [
          "users"
        ]
      }
    },
    "/ingredient-sets": {
      "post": {
        "operationId": "IngredientsController_createIngredientSet",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIngredientDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ingredient set created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientSetResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a new ingredient set",
        "tags": [
          "ingredient-sets"
        ]
      },
      "get": {
        "operationId": "IngredientsController_getUserIngredientSets",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of ingredient sets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IngredientSetResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get ingredient sets by user ID",
        "tags": [
          "ingredient-sets"
        ]
      }
    },
    "/ingredient-sets/{id}": {
      "patch": {
        "operationId": "IngredientsController_updateIngredientSet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIngredientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ingredient set updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientSetResponseDto"
                }
              }
            }
          }
        },
        "summary": "Update an ingredient set",
        "tags": [
          "ingredient-sets"
        ]
      }
    },
    "/recipes": {
      "post": {
        "operationId": "RecipeController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecipeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recipe saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeResponseDto"
                }
              }
            }
          }
        },
        "summary": "Save a recipe",
        "tags": [
          "recipes"
        ]
      },
      "get": {
        "operationId": "RecipeController_findByUserId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "favoritesOnly",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of recipes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedRecipeResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get recipes by user ID",
        "tags": [
          "recipes"
        ]
      }
    },
    "/recipes/{id}/favorite": {
      "patch": {
        "operationId": "RecipeController_toggleFavorite",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Favorite toggled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeResponseDto"
                }
              }
            }
          }
        },
        "summary": "Toggle recipe favorite status",
        "tags": [
          "recipes"
        ]
      }
    },
    "/recipes/{id}": {
      "delete": {
        "operationId": "RecipeController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recipe deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeResponseDto"
                }
              }
            }
          }
        },
        "summary": "Delete a recipe",
        "tags": [
          "recipes"
        ]
      }
    }
  },
  "info": {
    "title": "ThreeJS API",
    "description": "ThreeJS API documentation",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          }
        },
        "required": [
          "id",
          "email"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "password1",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ValidateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "password1"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "DeleteUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "id"
        ]
      },
      "CreateIngredientDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "example": 1
          },
          "ingredients": {
            "example": [
              "tomato",
              "onion",
              "garlic"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "userId",
          "ingredients"
        ]
      },
      "IngredientSetResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "ingredients": {
            "example": [
              "tomato",
              "onion",
              "garlic"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "userId": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "id",
          "ingredients",
          "createdAt",
          "userId"
        ]
      },
      "UpdateIngredientDto": {
        "type": "object",
        "properties": {
          "ingredients": {
            "example": [
              "tomato",
              "onion",
              "garlic"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "ingredients"
        ]
      },
      "CreateRecipeDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "トマトパスタ"
          },
          "content": {
            "type": "string",
            "example": "## トマトパスタの作り方\n..."
          },
          "ingredients": {
            "example": [
              "トマト",
              "玉ねぎ",
              "にんにく"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "servings": {
            "type": "number",
            "example": 2
          },
          "youtubeVideos": {
            "type": "object",
            "example": [
              {
                "videoId": "abc",
                "title": "Recipe video",
                "channelTitle": "Channel",
                "thumbnail": "https://..."
              }
            ],
            "nullable": true
          }
        },
        "required": [
          "userId",
          "name",
          "content",
          "ingredients",
          "servings"
        ]
      },
      "RecipeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "トマトパスタ"
          },
          "content": {
            "type": "string",
            "example": "## トマトパスタの作り方\n..."
          },
          "ingredients": {
            "example": [
              "トマト",
              "玉ねぎ",
              "にんにく"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "servings": {
            "type": "number",
            "example": 2
          },
          "youtubeVideos": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "isFavorite": {
            "type": "boolean",
            "example": false
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "userId": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "id",
          "name",
          "content",
          "ingredients",
          "servings",
          "isFavorite",
          "createdAt",
          "userId"
        ]
      },
      "PaginatedRecipeResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeResponseDto"
            }
          },
          "total": {
            "type": "number",
            "example": 10
          }
        },
        "required": [
          "items",
          "total"
        ]
      }
    }
  }
}