# apps/api/src/users/dto/validate-user.dto.ts
import { ApiProperty } from '@nestjs/swagger';

export class ValidateUserDto {
  @ApiProperty({ example: 'user@example.com' })
  email: string;

  @ApiProperty({ example: 'password1' })
  password: string;
}