# apps/api/src/users/users.controller.ts
import {
  Controller,
  Get,
  Post,
  Body,
  HttpException,
  HttpStatus,
  UsePipes,
  ValidationPipe,
  Delete,
  Logger,
} from '@nestjs/common';
import { ApiTags, ApiOperation, ApiResponse } from '@nestjs/swagger';
import { UsersService } from './users.service';
import { CreateUserDto } from './dto/create-user.dto';
import { UserResponseDto } from './dto/user-response.dto';
import { ValidateUserDto } from './dto/validate-user.dto';
import { DeleteUserDto } from './dto/delete-user.dto';
import { z } from 'zod';
import { ZodDto } from 'nestjs-zod';

const CreateUserInput = z.object({
  email: z.string().email(),
  password: z
    .string()
    .min(8, 'Password must be at least 8 characters long')
    .regex(
      /^(?=.*[a-zA-Z])(?=.*\d)[A-Za-z\d]{8,}$/,
      'Password must contain both letters and numbers',
    ),
});

// Simplicity - Just a simple controller to create and list users
@ApiTags('users')
@Controller('users')
export class UsersController {
  private readonly logger = new Logger(UsersController.name);

  constructor(private readonly usersService: UsersService) {}

  @Get()
  @ApiOperation({ summary: 'Get all users' })
  @ApiResponse({ status: 200, description: 'List of users', type: [UserResponseDto] })
  getUsers() {
    return this.usersService.listUsers();
  }

  @Post()
  @ApiOperation({ summary: 'Create a new user' })
  @ApiResponse({ status: 201, description: 'User created', type: UserResponseDto })
  @ApiResponse({ status: 409, description: 'User already exists' })
  async createUser(@Body() user: CreateUserDto) {
    this.logger.log(`Creating user with email: ${user.email}`);

    try {
      const createdUser = await this.usersService.createUser(
        user.email,
        user.password,
      );
      this.logger.log(`User created successfully: ${createdUser.id}`);

      // パスワードを除外してレスポンスを返す
      const { password, ...userWithoutPassword } = createdUser;
      return userWithoutPassword;
    } catch (error) {
      this.logger.error(
        `Error creating user: ${error instanceof Error ? error.message : 'Unknown error'}`,
      );

      if (
        error instanceof Error &&
        error.message === 'User with this email already exists'
      ) {
        throw new HttpException(
          'User with this email already exists',
          HttpStatus.CONFLICT,
        );
      }
      console.error('Create user error:', error);
      throw new HttpException(
        'Failed to create user',
        HttpStatus.INTERNAL_SERVER_ERROR,
      );
    }
  }

  @Post('validate')
  @ApiOperation({ summary: 'Validate user credentials' })
  @ApiResponse({ status: 200, description: 'User validated', type: UserResponseDto })
  @ApiResponse({ status: 401, description: 'Invalid credentials' })
  async validate(@Body() body: ValidateUserDto) {
    const user = await this.usersService.validateUser(
      body.email,
      body.password,
    );
    if (!user) {
      throw new HttpException('Invalid credentials', HttpStatus.UNAUTHORIZED);
    }
    return user;
  }

  @Delete()
  @ApiOperation({ summary: 'Delete a user' })
  @ApiResponse({ status: 200, description: 'User deleted' })
  @ApiResponse({ status: 404, description: 'User not found' })
  deleteUser(@Body() body: DeleteUserDto) {
    return this.usersService.deleteUser(body.id);
  }
}